requirements.txt
pandas<2.0.0
plotly==5.18.0
streamlit==1.30.0
joblib==1.3.2
google-cloud-bigquery==3.16.0
google-auth == 2.26.2
db-dtypes==1.2.0
prophet==1.1.5